% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFasToVirtualSdy.R
\name{addAnnoToVirtualSdy}
\alias{addAnnoToVirtualSdy}
\title{Function to migrate a copy of FA / FAS from /Studies/ to a virtual study container}
\usage{
addAnnoToVirtualSdy(ISserver, virtualSdy)
}
\arguments{
\item{ISserver}{immunespace server, either test or prod}

\item{virtualSdy}{subdirectory within HIPC container}
}
\description{
Function to migrate a copy of FA / FAS from /Studies/ to a virtual study container
}
