% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyMatricesToVirtualSdy.R
\name{copyMatricesToVirtualSdy}
\alias{copyMatricesToVirtualSdy}
\title{Function to migrate a copy of EMs from /Studies/ to a virtual study container}
\usage{
copyMatricesToVirtualSdy(ISserver, virtualSdy)
}
\arguments{
\item{ISserver}{immunespace server, either test or prod}

\item{virtualSdy}{subdirectory within HIPC container}
}
\description{
Function to migrate a copy of EMs from /Studies/ to a virtual study container
}
