% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgncAlias2Symbol}
\alias{hgncAlias2Symbol}
\title{HGNC alias2symbol mapping table}
\format{
data.table
\describe{
  \item{SYMBOL}{Gene Symbol}
  \item{ALIAS}{Previous or current gene aliases}
  ...
}
}
\usage{
hgncAlias2Symbol
}
\description{
data.table holding mapping of gene aliases to symbol
}
\keyword{datasets}
