% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgncAlias2Symbol_version}
\alias{hgncAlias2Symbol_version}
\title{HGNC dataset version}
\format{
string
\describe{
  \item{value}{date when HGNC dataset was downloaded and used in updates}
  ...
}
}
\usage{
hgncAlias2Symbol_version
}
\description{
Date HGNC dataset was used to update gene symbols. Pushed
to microarray.FeatureAnnotationSet.
}
\keyword{datasets}
