% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnoDf.R
\name{makeAnnoDf}
\alias{makeAnnoDf}
\title{Function to map probe ids to gene symbols from an annotation package}
\usage{
makeAnnoDf(annoPkg, outPath = NULL)
}
\arguments{
\item{outPath}{file path for tsv to write out for upload to ImmuneSpace}

\item{probes}{character vector of probe ids}

\item{annoPks}{name of annotation package to use}
}
\description{
Function to map probe ids to gene symbols from an annotation package
}
