% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orgHsEgDb_version}
\alias{orgHsEgDb_version}
\title{org.Hs.eg.db packageVersion}
\format{string
\describe{
  \item{value}{packageVersion of org.Hs.eg.db used in updates}
  ...
}}
\usage{
orgHsEgDb_version
}
\description{
packageVersion of org.Hs.eg.db used to update gene symbols. Pushed
to microarray.FeatureAnnotationSet.
}
\keyword{datasets}
