% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUpdatedBTM.R
\name{outputUpdatedBTM}
\alias{outputUpdatedBTM}
\title{Function to output a gmt file of the latest updated btm data from library}
\usage{
outputUpdatedBTM()
}
\description{
Function to output a gmt file of the latest updated btm data from library
}
