% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateAnnoVec.R
\name{updateAnno}
\alias{updateAnno}
\title{Function to update annotation of a character vector of gene aliases}
\usage{
updateAnno(vec)
}
\arguments{
\item{vec}{character vector of gene aliases}
}
\description{
Function to update annotation of a character vector of gene aliases
}
